+++ pkg/contracts/testdappv2/TestDAppV2.sol
@@ -167,6 +167,7 @@ contract TestDAppV2 {
 
     // Revertable interface
     function onRevert(RevertContext calldata revertContext) external {
+        require(!isRevertMessage(string(revertContext.revertMessage)));
 
         // if the chain is ZetaChain, consume gas to test the gas consumption
         // we do it specifically for ZetaChain to test the outbound processing workflow