+++ changelog.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+* [3414](https://github.com/zeta-chain/node/pull/3414) - support advanced abort workflow (onAbort)
 * [3461](https://github.com/zeta-chain/node/pull/3461) - add new `ConfirmationParams` field to chain params to enable multiple confirmation count values, deprecating `confirmation_count`
 * [3489](https://github.com/zeta-chain/node/pull/3489) - add Sui chain info
 * [3455](https://github.com/zeta-chain/node/pull/3455) - add `track-cctx` command to zetatools
@@ -11,6 +12,9 @@
 * [3469](https://github.com/zeta-chain/node/pull/3469) - add `MsgRemoveInboundTracker` to remove inbound trackers. This message can be triggered by the emergency policy.
 * [3450](https://github.com/zeta-chain/node/pull/3450) - SOL withdraw and call integration
 * [3520] (https://github.com/zeta-chain/node/pull/3520) - SPL withdraw and call integration
+* [3534] (https://github.com/zeta-chain/node/pull/3534) - Add Sui deposit & depositAndCall
+* [3541](https://github.com/zeta-chain/node/pull/3541) - implement `MsgUpdateZRC20Name` to update the name or symbol of a ZRC20 token
+* [3520](https://github.com/zeta-chain/node/pull/3520) - SPL withdraw and call integration
 
 ### Refactor